apps/api/pyproject.toml
[project]
name = "terrainpilot-api"
version = "0.1.0"
description = "FastAPI backend for TerrainPilot"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.15.1",
  "argon2-cffi>=23.1.0",
  "asyncpg>=0.30.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.11",
  "greenlet>=3.1.1",
  "httpx>=0.28.1",
  "itsdangerous>=2.2.0",
  "pydantic-settings>=2.8.1",
  "pyjwt>=2.10.1",
  "redis>=5.2.1",
  "sqlalchemy>=2.0.39",
  "uvicorn>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "aiosqlite>=0.21.0",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
